LayoutTests/imported/w3c/web-platform-tests/css/selectors/invalidation/has-side-effect-expected.txt


PASS Initial colors
FAIL Matches after #blocks_match removed assert_equals: expected "rgb(255, 0, 0)" but got "rgb(128, 128, 128)"
PASS Does not match after #blocks_match added
